requirements.txt
Pillow~=9.4.0
dearpygui~=1.8.0
